tts_service/requirements.txt
coqui-tts>=0.20.0
fastapi==0.104.1
uvicorn==0.24.0
soundfile==0.12.1
python-multipart==0.0.6